scripts/requirements.txt
aggregate6==1.0.12
arouteserver==1.21.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==21.3
py-radix==0.10.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.31.0
urllib3==2.0.3
